Fix visibility of focused skip link on docs site (#2405)
* Fix visibility of focused skip link on docs site

The `sr-only` utility class used on the “Skip to Content” link at the top of each page in the docs is designed to be paired with the `focus:not-sr-only` class. `focus:not-sr-only` was missing on the skip link, meaning it was never shown even when focused.

* Fix visibility of focused skip link in docs example project

Apply the same fix as ac890b56 to the same component in the docs starter project.